+++ package.json
@@ -46,7 +46,7 @@
     "fuse-box-test-using-old-version": "^1.0.0",
     "fuse-box-testcase1": "^1.0.0",
     "fuse-box-testcase3": "^1.0.3",
-    "fuse-test-runner": "^1.0.9",
+    "fuse-test-runner": "^1.0.10",
     "gulp": "3.9.1",
     "gulp-bump": "2.4.0",
     "gulp-changelog-generator": "git+https://github.com/fuse-box/gulp-changelog-generator.git",
